fix: responsing via webhooks missed header.
Telegram only processes methods returned as json to a webhook call if
the Content-Type header is set to "application/json".

This did not happen with express and koa, as grammy just wrote the json
stringified to the body instead of directly passing the json object,
causing the frameworks to not set the json Content-Type header.

This fix is only for express-like and koa frameworks, others might still
have this bug.