test: only test types of adapters
In some cases, we fake the requests, too. This is only done when surrounding frameworks would be tested by calling the webhook adapter directly with the respective request objects. In other cases, such as when testing via `app.post("/", handler)`, we only create the adapter and pass it to the framework, without actually simulating a request.

We may deicde to
  1) skip supplying fake update in the future, such as by wrapping the code in a lambda function, or
  2) to fake requests for all frameworks and actually test the runtime code, too.